new_setup branch merge with trunk.
If any of you are using this, you will need to update your
 import Box2D2 
to 
 import Box2D

and follow another couple updates with buoyancy/etc.

+++ trunk/Box2D/Dynamics/Contacts/b2PolyAndEdgeContact.cpp
@@ -278,7 +278,7 @@ void b2PolyAndEdgeContact::b2CollidePolyAndEdge(b2Manifold* manifold,
 			manifold->normal = b2Mul(xf1.R, normals[separationIndex]);
 			manifold->points[0].separation = separationMax;
 			manifold->points[0].id.features.incidentEdge = (uint8)separationIndex;
-			manifold->points[0].id.features.incidentVertex = (uint8)b2_nullFeature;
+			manifold->points[0].id.features.incidentVertex = b2_nullFeature;
 			manifold->points[0].id.features.referenceEdge = 0;
 			manifold->points[0].id.features.flip = 0;
 			if (separationV1) {
